requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.13
langgraph==0.0.20
chromadb==0.4.22
sentence-transformers==2.3.1

# PDF processing
pypdf==3.17.4
pypdf2==3.0.1

# API and UI
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional LLM integrations (uncomment as needed)
langchain-groq==0.0.10
# openai==1.10.0
# groq==0.4.1
# langchain-openai==0.0.5
